+++ apps/web/pages/_app.tsx
@@ -5,7 +5,7 @@ import { AppProps } from 'next/app'
 import { WithRouterProps } from 'next/dist/client/with-router'
 import React from 'react'
 import { RecoilRoot } from 'recoil'
-import { LayoutFactory } from '../src/layout/Layout'
+import { LayoutFactory } from '../src/templates/Layout'
 import { LoginUserModal, RegisterUserModal } from '../src/useCases/user'
 import { Page, useApollo } from '@codelab/frontend'
 import './App.scss'
