+++ apps/web/pages/index.tsx
@@ -1,8 +1,8 @@
 import { NextPage } from 'next'
 import React from 'react'
-import { HomeClients } from '../src/home/Home-clients'
-import { HomeFeatures } from '../src/home/Home-features'
-import { HomeJumbo } from '../src/home/Home-jumbo'
+import { HomeClients } from '../src/pages/home/Home-clients'
+import { HomeFeatures } from '../src/pages/home/Home-features'
+import { HomeJumbo } from '../src/pages/home/Home-jumbo'
 import { withAuthServerSideProps } from '@codelab/frontend'
 
 const HomePage: NextPage = (props) => {
