+++ apps/web/src/pages/builder/Builder.tsx
@@ -1,13 +1,13 @@
 import { Layout } from 'antd'
 import React, { PropsWithChildren, useContext } from 'react'
 import { useRecoilState } from 'recoil'
-import { LayoutPaneVisibility, useLayout } from '../layout/layout-state'
-import { AppContext, AppProvider } from '../useCases/apps/AppProvider'
+import { LayoutPaneVisibility, useLayout } from '../../templates/layout-state'
+import { AppContext, AppProvider } from '../../useCases/apps/AppProvider'
+import { componentItemState } from '../../useCases/component/paneMain/Component-item--state'
 import { BuilderPaneController } from './Builder-pane-controller'
 import { PaneConfig } from './pane-config/Pane-config'
 import { BuilderDetails } from './pane-details/Pane-details'
 import { PaneMain } from './pane-main/Pane-main'
-import { componentItemState } from './pane-main/component/Component-item--state'
 import { BuilderTabSidebar } from './tabs-sidebar/Tabs-sidebar'
 import { contentStyle } from '@codelab/frontend'
