+++ apps/web/src/pages/builder/pane-details/Pane-details.tsx
@@ -1,8 +1,8 @@
 import React from 'react'
-import { LayoutTab } from '../../layout/layout-state'
-import { PageContainerCreate } from '../../useCases/pages/Page-container--create'
-import { PageContainerUpdateDelete } from '../../useCases/pages/Page-container--updateDelete'
-import { usePage } from '../../useCases/pages/usePage'
+import { LayoutTab } from '../../../templates/layout-state'
+import { PageContainerCreate } from '../../../useCases/pages/Page-container--create'
+import { PageContainerUpdateDelete } from '../../../useCases/pages/Page-container--updateDelete'
+import { usePage } from '../../../useCases/pages/usePage'
 import { BuilderPaneController } from '../Builder-pane-controller'
 
 export const BuilderDetails = () => {
