+++ apps/web/src/templates/Layout.tsx
@@ -1,8 +1,8 @@
 import { Editor } from '@craftjs/core'
 import { WithRouterProps } from 'next/dist/client/with-router'
 import React, { PropsWithChildren } from 'react'
-import { Builder } from '../builder/Builder'
-import { HomeLayout } from '../home'
+import { Builder } from '../pages/builder/Builder'
+import { HomeLayout } from '../pages/home'
 import { AppProvider } from '../useCases/apps/AppProvider'
 import { Dashboard } from './Dashboard'
 import { PageType, elementTypeMap } from '@codelab/frontend'
