+++ apps/web/src/useCases/component/paneMain/Component-item--state.ts
