--- apps/web/src/useCases/functions/createFunction/CreateFunctionForm.tsx
@@ -1,3 +1,0 @@
-export const CreateFunctionForm = () => {
-  return null
-}
