+++ apps/web/src/useCases/lambda/createLambda/CreateLambdaForm.tsx
@@ -1,0 +1,3 @@
+export const CreateLambdaForm = () => {
+  return null
+}
