+++ apps/web/src/useCases/pages/paneMain/Page-main--close.tsx
@@ -1,6 +1,6 @@
 import { CloseOutlined } from '@ant-design/icons'
 import React from 'react'
-import { LayoutPaneVisibility } from 'apps/web/src/layout/layout-state'
+import { LayoutPaneVisibility } from 'apps/web/src/templates/layout-state'
 import { usePage } from 'apps/web/src/useCases/pages/usePage'
 
 export const PageMainClose = () => {
