+++ apps/web/src/useCases/style/paneConfig/Pane-config--style.tsx
