+++ libs/frontend/src/renderer/Renderer-components.tsx
@@ -1,8 +1,8 @@
 import { SerializedNodes, useEditor } from '@craftjs/core'
 import React, { useContext, useRef } from 'react'
 import { useRecoilState } from 'recoil'
-import { paneConfigState } from '../../../../apps/web/src/builder/pane-config/Pane-config'
-import { PaneConfigHandlersProps } from '../../../../apps/web/src/builder/pane-config/Pane-config--handlers'
+import { paneConfigState } from '../../../../apps/web/src/pages/builder/pane-config/Pane-config'
+import { PaneConfigHandlersProps } from '../../../../apps/web/src/pages/builder/pane-config/Pane-config--handlers'
 import { NodeA } from '../../../modules/graph/src/core/domain/node/Node'
 import { useOverlayToolbar } from '../components'
 import { CLICK_OVERLAY_ID } from './Overlay-click'
